Add support for reading the git config
Currently there's no way to work out what a config variable should be
with git simply by reading text files. This is because the git exec
reads from different places depending on what prefix was during build.

This will allow us to use the binary to resolve that information.

This is important because there is a `core.commentChar` configuration
variable which affects how commit messages should be parsed.