Allow ignoring comment lines in commits from files
When git is creating a commit, the message it reads from a file does
not include the lines beginning with a comment character. This means
that the commit files that you get passed with a git commit message
hook is not exactly what will appear in git if saved.

The comment character in versions of git later than 1.8.2 is
configurable, in versions prior to that it's always '#'.

You might however notice that the CommitMessage defaults the character
to null. This is because a commit message from the repository may have
had any comment character set when it was created, so you could see
lines in a commit message from the repository that begin with what is
now the comment character.

The lines with the comments can still be accessed via getLines methods,
however the semantic methods of getBody and getSubject do not return
them.