Stop config::has() throwing exception on missing
This function is a helper to make it nicer to check if a configuration
key exists. It's supposed to return true or false based on if the
config key exists or not.

When git tells you if a config key does not exist it has a return code
of `1`. This is a failure status code, so we throw an exception, we
need to handle that and turn it into something more friendly.